refactor: split `panel.py` into multiple modules
- Improve modularity and logic separation
- Facilitate ease of expansion and maintenance
