+++ portal/ui/globals.py
@@ -1,0 +1,1 @@
+MODAL_OPERATORS = {}
